+++ src/tr1/game/console/cmd/easy_config.c
@@ -14,6 +14,7 @@ static COMMAND_TO_OPTION_MAP m_CommandToOptionMap[] = {
     { "vsync", &g_Config.rendering.enable_vsync },
     { "wireframe", &g_Config.rendering.enable_wireframe },
     { "fps", &g_Config.rendering.fps },
+    { "lighting", &g_Config.rendering.enable_lighting },
     { nullptr, nullptr },
 };
 
@@ -40,3 +41,4 @@ REGISTER_CONSOLE_COMMAND("vsync", M_Entrypoint, GS_ID(CONSOLE_HELP_VSYNC))
 REGISTER_CONSOLE_COMMAND(
     "wireframe", M_Entrypoint, GS_ID(CONSOLE_HELP_WIREFRAME))
 REGISTER_CONSOLE_COMMAND("fps", M_Entrypoint, GS_ID(CONSOLE_HELP_FPS))
+REGISTER_CONSOLE_COMMAND("lighting", M_Entrypoint, GS_ID(CONSOLE_HELP_LIGHTING))